src/requirements.txt
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
contextlib2==21.6.0
idna==3.3
jsonschema==4.9.0
pyrsistent==0.18.1
PyYAML==6.0
requests==2.28.1
schema==0.7.5
urllib3==1.26.11
